test(editor): harden VS Code host integration
